refactor: implement new architecture PoC #15
