+++ example/index.html
@@ -8,6 +8,6 @@
 		<link rel="stylesheet" href="build/bundle.css" />
 	</head>
 	<body>
-		<script defer src="build/index.js"></script>
+		<script type="module" defer src="build/index.js"></script>
 	</body>
 </html>
