+++ package.json
@@ -32,9 +32,7 @@
 		"cross-env": "^7.0.2",
 		"husky": "^4.2.5",
 		"prettier": "^2.1.1",
-		"prettier-plugin-svelte": "^1.2.1"
-	},
-	"dependencies": {
+		"prettier-plugin-svelte": "^1.2.1",
 		"svelte": "^3.24.1"
 	}
 }
